Fixed several mistakes in the example unit test: -
a) incorrect positioning of a collection subrange,
b) transposed arguments to a call to calculate the number of combinations to choose from,
c) the 'NoOperation' enumeration wasn't creating an operation - this was deliberate but messed up the subsequent logic for placing operations.

Also sorted out a problem to do with the type constraint for test *levels* infecting the types of the computed test cases, manifest only at run time.

Finally, there was an overflow problem in the RandomBehaviour class.

Still got bugs at this point in the test case generation implementation, but making a waypoint commit.

--HG--
extra : convert_revision : svn%3Ab091aa91-a314-92f1-5d41-19d272613d48/trunk%40129