Extend test to check that a synthesizing factory with permutations doesn't miss any combination of levels.
This currently breaks, as expected - the test can only be met if the
strength guarantee is raised locally for each synthesizing node that
creates permutations. Currently the implementation slots the implicit
test variable for permutation levels into the shared overall strength,
thus being deficient (and hopefully the cause of said expected failure).