Fix bug in test checking permutation via synthesizing factories.
Had glibly assumed that not setting the flag to generate a permutation
example in a tree of factories would generate the same tree structure as
if the flag had been set - this introduced a bug in the previous commit.
Have corrected that, although the test still fails - but for the right
reasons this time!