Generate fewer permutation examples for unit test that is just including them for the sake of completeness.
The permutation support is explicitly tested elsewhere - this is just to
make sure that another test has checked the full variety of
possibilities. As that test was taking an inordinately long time to
complete (because the combination strength for test cases involving
permutations has just gone up due to the previous commit), this commit
was done to fix that.